fix(cloze): move import.meta.glob into a module so node -c passes
CIの 'Check JavaScript syntax' (node -c script.js) は CommonJS として構文チェックするため、script.js 内の import.meta.glob が 'Cannot use import.meta outside a module' で失敗していた。glob と mergePhraseCollections を src/data/phrase-collections.js (ESモジュール) に分離し、script.js は動的 import 経由で利用する。node -c script.js / build / content-test 通過を確認。
